Draft of Austin City Council Resolution for Crypto Feasibility Study
This was my version of the resolution that ultimately did not go before the Austin City Council. However, I still think that this version would be better since it explicitly guides and directs city employees in the right direction rather than them trying to go out and figure things out on their own.